packages/vision/requirements.txt
numpy==1.17.0
opencv-contrib-python-headless==4.1.0.25
pyzmq==18.1.0
